+++ src/service_inspectors/s7comm/ips_s7comm_error_class.cc
@@ -30,48 +30,13 @@
 #include "profiler/profiler.h"
 
 #include "s7comm.h"
+#include "s7comm_helpers.h"
+
 #include <iostream>
 
 using namespace snort;
 
 static const char* s_name = "s7comm_error_class";
-
-//-------------------------------------------------------------------------
-// func lookup
-//-------------------------------------------------------------------------
-
-struct S7commFuncMap
-{
-    const char* name;
-    uint8_t func;
-};
-
-/* Mapping of name -> message type for 's7comm_func' option. */
-static S7commFuncMap s7comm_func_map[] =
-{
-    { "job_request",    0x01 },
-    { "ack",            0x02 },
-    { "ack_data",       0x03 },
-    { "userdata",       0x07 }
-};
-
-
-struct S7commErrorClassMap
-{
-    const char* name;
-    uint8_t error_class;
-};
-
-static S7commErrorClassMap s7comm_error_class_map[] =
-{
-    { "no_error",                        0x00 },
-    { "application_relationship_error",  0x81 },
-    { "object_definition_error",         0x82 },
-    { "no_resources_available_error",    0x83 },
-    { "error_on_service_processing",     0x84 },
-    { "error_on_supplies",               0x85 },
-    { "access_error",                    0x87 }
-};
 
 static bool get_error_class(const char* s, long& n)
 {
