+++ src/service_inspectors/s7comm/ips_s7comm_item_count.cc
@@ -39,25 +39,6 @@ using namespace snort;
 static const char* s_name = "s7comm_item_count";
 
 //-------------------------------------------------------------------------
-// func lookup
-//-------------------------------------------------------------------------
-
-struct S7commFuncMap
-{
-    const char* name;
-    uint8_t func;
-};
-
-/* Mapping of name -> message type for 's7comm_func' option. */
-static S7commFuncMap s7comm_func_map[] =
-{
-    { "job_request",    0x01 },
-    { "ack",            0x02 },
-    { "ack_data",       0x03 },
-    { "userdata",       0x07 }
-};
-
-//-------------------------------------------------------------------------
 // item_count option
 //-------------------------------------------------------------------------
